+++ tools/contract_artifact_builder.py
@@ -8,11 +8,13 @@
 """
 
 import argparse
-import base64
-import json
 import gzip
-import re
+import json
 import os
+import re
+
+DEFAULT_ARTICLE_LIBRARY = 'artifacts.json.gz'
+DEFAULT_COMMAND = 'create'
 
 
 def import_artifact_file(filename, data):
@@ -37,11 +39,12 @@ def import_artifact_file(filename, data):
     data[network_name][contract_name] = data_json
     print(f'Adding {network_name}/{contract_name}')
 
+
 def outptut_as_python_file(filename, data):
     with open(filename, 'w') as fp:
         fp.write('# auto generated artifle files\n')
         fp.write('artifact_data = {\n',)
-        for network_name, item in artifact_data.items():
+        for network_name, item in data.items():
             fp.write(f"    '{network_name}': {{ \n")
             for contract_name, abi_item in item.items():
                 fp.write(f"        '{contract_name}': {{\n")
@@ -50,6 +53,7 @@ def outptut_as_python_file(filename, data):
 
             fp.write('    }\n')
         fp.write('\n')
+
 
 def output_as_json_file(filename, data, is_compressed):
     if is_compressed:
@@ -60,8 +64,14 @@ def output_as_json_file(filename, data, is_compressed):
             json.dump(data, fp)
 
 
-def main():
+def list_artifact_data(data):
+    print(f'Network Name         Contract Name')
+    for network_name, contract_names in data.items():
+        for contract_name, item in contract_names.items():
+            print(f'{network_name:20} {contract_name}')
 
+
+def main():
 
     parser = argparse.ArgumentParser(description='Contract Artifact Library Builder')
 
@@ -80,45 +90,67 @@ def main():
     )
 
     parser.add_argument(
-        'output_filename',
-        metavar='out-file',
-        help='main artifact library filename to create/update'
+        '-f',
+        '--file',
+        default=DEFAULT_ARTICLE_LIBRARY,
+        help=f'filename of the artifact library filename to manage. Default: {DEFAULT_ARTICLE_LIBRARY}'
+    )
+
+    parser.add_argument(
+        'command',
+        default=DEFAULT_COMMAND,
+        help='''command to execute, can be the following
+
+            create or c - create a new artifact library. Default option
+            udpate or u - update a current library
+            list or l   - list the contents of a library
+
+            '''
     )
 
     parser.add_argument(
         'artifact_filename',
         metavar='artifact-file',
-        nargs='+',
+        nargs='?',
         help='input abi filename'
     )
     args = parser.parse_args()
 
-    output_file = args.output_filename
-    if re.search('\.gz$', output_file):
-        output_file_gz = output_file
-        output_file = re.sub('\.gz$', '', output_file_gz)
+    library_file = args.file
+    if re.search(r'\.gz$', library_file):
+        library_file_gz = library_file
+        library_file = re.sub(r'\.gz$', '', library_file_gz)
     else:
-        output_file_gz = f'{output_file}.gz'
+        library_file_gz = f'{library_file}.gz'
 
     artifact_data = {}
 
-    if os.path.exists(output_file) or os.path.exists(output_file_gz):
-        if not args.update:
-            print(f'The output file {output_file} already exists, please use the -u --update option to overwite this file')
+    command_char = args.command.lower()[0]
+
+    if os.path.exists(library_file) or os.path.exists(library_file_gz):
+        if command_char == 'c':
+            print(f'You cannot create an existing library file "{library_file}"')
             return
 
-        if os.path.exists(output_file_gz):
-            with gzip.open(output_file_gz, 'rt') as fp:
+        if os.path.exists(library_file_gz):
+            with gzip.open(library_file_gz, 'rt') as fp:
                 artifact_data = json.load(fp)
 
-        if os.path.exists(output_file):
-            with open(output_file, 'r') as fp:
+        if os.path.exists(library_file):
+            with open(library_file, 'r') as fp:
                 artifact_data = json.load(fp)
 
-    import_artifact_file(args.artifact_filename, artifact_data)
+    if command_char == 'c' or command_char == 'u':
+        if args.artifact_filename:
+            import_artifact_file(args.artifact_filename, artifact_data)
+            output_as_json_file(library_file_gz, artifact_data, True)
+        else:
+            print('Please provide a single or multiple contract article files to create/udpate to this library')
 
-    output_as_json_file(output_file_gz, artifact_data, True)
-
+    elif command_char == 'l':
+        list_artifact_data(artifact_data)
+    else:
+        print(f'Unknown command {args.command}')
 
 
 if __name__ == "__main__":